dpkg-maintscript-helper: Support moving a conffile not being shipped
When the conffile stops being shipped, it might make sense to move it
somewhere else to preserve it. Do not fail if the new conffile does not
exist.

Closes: #767003

+++ scripts/dpkg-maintscript-helper.sh
@@ -210,7 +210,9 @@ finish_mv_conffile() {
 	ensure_package_owns_file "$PACKAGE" "$OLDCONFFILE" || return 0
 
 	echo "Preserving user changes to $NEWCONFFILE (renamed from $OLDCONFFILE)..."
-	mv -f "$NEWCONFFILE" "$NEWCONFFILE.dpkg-new"
+	if [ -e "$NEWCONFFILE" ]; then
+		mv -f "$NEWCONFFILE" "$NEWCONFFILE.dpkg-new"
+	fi
 	mv -f "$OLDCONFFILE" "$NEWCONFFILE"
 }
